Object_Detect/requirements.txt
Django
Pillow
gunicorn
tensorflow
WhiteNoise
opencv-python
numpy